requirements.txt
faiss-cpu>=1.8.0
numpy>=1.26.0
ollama>=0.3.0
tqdm>=4.66.0

